[hub] Make overlay not touchable while soft-hidden
When there are no devices, the overlay soft-hides. (It does not set visibility to gone to simplify state
management, though we could easily revisit that.) However, it was still blocking touches. Set it to not
touchable while there are no devices.

Also uses window param alpha instead of Flutter alpha.